замена int на float

+++ kalkulation_core.py
@@ -28,8 +28,8 @@ class MyKalkulationCore(QtCore.QThread):
         print(self.glubina)
         print(self.material)
         print(self.material_rate)
-        skorost = int(self.ploshad) * 1.8  # 1.8 - стандартное время обработки 1мм2
-        skorostsloev = int(self.glubina) * skorost  # кол-во слоев * Nмм2/сек
+        skorost = float(self.ploshad) * 1.8  # 1.8 - стандартное время обработки 1мм2
+        skorostsloev = float(self.glubina) * skorost  # кол-во слоев * Nмм2/сек
         resultTrud = round(skorostsloev / 3600, 3)  # трудоемкость = площадь обработки*кол-во секунд / 3600(перевод в н/ч)
         """Вставить поиск цены по сравнению столбца material_mark из БД material_cost с LEMaterial"""
         """myinput = main.KalkulationWindow.setLineEditMaterial(text)
@@ -41,10 +41,10 @@ class MyKalkulationCore(QtCore.QThread):
         #     work_cost_.append(row)
         # print(f"выбран материала" + {work_cost_} + ".")
 """
-        materialcost = int(self.material_rate) * 655  # вставить вместо 655 (цена за кг) данные из БД material_cost через условие
+        materialcost = float(self.material_rate) * 655  # вставить вместо 655 (цена за кг) данные из БД material_cost через условие
         print(f"Стоимость материала {materialcost}")
         resultSebest = round(resultTrud * 2350 + materialcost, 2)  # 2350 =  цена нормочаса в руб. с НДС
-        vremyapartii = skorostsloev * int(self.kolichestvo) // (3600*7.5)
+        vremyapartii = skorostsloev * float(self.kolichestvo) // (3600*7.5)
         print(f"Трудоёмкость {resultTrud}")
         print(f"Себестоимость {resultSebest}")
         print(f"Время обработки партии {vremyapartii} дней")
